ok启动条页面中调整启动条的宽度

+++ Core/Models/AppConfig.cs
@@ -51,7 +51,7 @@ namespace PromptMasterv5.Core.Models
         private bool enableLaunchBar = true;
 
         [ObservableProperty]
-        private int launchBarWidth = 6;
+        private double launchBarWidth = 6.0;
 
         [ObservableProperty]
         private string launchBarHotkey = "Alt+L";
